Extract common functionality into AbstractDateConverter
