+++ graylog2-server/src/main/java/org/graylog2/inputs/converters/AbstractDateConverter.java
@@ -1,0 +1,56 @@
+/**
+ * This file is part of Graylog.
+ *
+ * Graylog is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Graylog is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.graylog2.inputs.converters;
+
+import org.graylog2.plugin.inputs.Converter;
+import org.joda.time.DateTimeZone;
+
+import java.util.Map;
+
+import static com.google.common.base.MoreObjects.firstNonNull;
+import static com.google.common.base.Strings.emptyToNull;
+
+public abstract class AbstractDateConverter extends Converter {
+    private static final DateTimeZone DEFAULT_TIME_ZONE = DateTimeZone.forID("Etc/UTC");
+
+    protected final DateTimeZone timeZone;
+
+    public AbstractDateConverter(Type type, Map<String, Object> config) {
+        super(type, config);
+
+        this.timeZone = buildTimeZone(config.get("time_zone"));
+    }
+
+    private static DateTimeZone buildTimeZone(Object timeZoneId) {
+        if (timeZoneId instanceof String) {
+            try {
+                final String timeZoneString = (String) timeZoneId;
+                final String zoneId = firstNonNull(emptyToNull(timeZoneString.trim()), "Etc/UTC");
+                return DateTimeZone.forID(zoneId);
+            } catch (IllegalArgumentException e) {
+                return DEFAULT_TIME_ZONE;
+            }
+        } else {
+            return DEFAULT_TIME_ZONE;
+        }
+    }
+
+    @Override
+    public boolean buildsMultipleFields() {
+        return false;
+    }
+}
